requirements.txt
tensorflow==2.12
flask
unidecode
flask-cors
pillow